Ensure core ID for topology.Node is correct
The root cause of the bug from issue #14 was the same issue from #11.
When processing the NUMA node core_id map from sysfs on Linux, we were
not appropriately removing the trailing newline from the []byte read
from ioutil.Readfile().

This fixes that issue and adds printing of core and cache information
for the TopologyInfo struct.

Closes Issue #14