fill out GraphicsCard.DeviceInfo field
Loops over all found graphics card addresses and fills out a
PCIDeviceInfo struct pointer for the card.

Modifies the PCIDeviceInfo struct by adding an Address field, which is
simply the PCI address of the device. After adding a String() method to
PCIDeviceInfo, we now print out the relevant class, vendor and product
name for the device.

Issue #4