lint-fixes: cpu module identifiers
Addresses a number of golint suggestions for the cpu module. Many of
these suggestions were around doc comments and changing Id -> ID. I've
added attributes called ID to the Processor and ProcessorCore structs,
but left the old Id attribute in the struct, with a note that these
attributes are deprecated in 0.2 and will be removed in 1.0.

Issue #51