output HostInfo as fully-contained JSON object
Before this patch, we were outputting the JSON strings for different
elements one after another. So, when executing `ghwc -f json` without
specifying a particular type of information to query (e.g. memory), we
were outputting invalid JSON, like this:

```
{
  "memory": ...
}
{
  "block": ...
}
```

This patch adds a YAMLString() and JSONString() method to the HostInfo
struct and modifies the `ghwc` command line tool to use the String(),
JSONString() or YAMLString() methods when outputting information about
the host (and all the subcomponents).

So now the JSON emitted is valid:

```
{
  "memory": ...,
  "block": ...,
  ...
}

Issue #78