correct calculation of NICCapability.CanChange
Fixes a buggy calculation of whether a NIC capability/feature can be
changed. There were two errors in the previous way of parsing the
feature line in the `ethtool -k` output.

The first error was that any line that did not have three parts was
considered non-changeable. This was incorrect, since this:

 scatter-gather: off

actually indicates that the feature can be enabled, and the line has two
fields (parts) not three.

The second error was that we were looking for the string "fixed" instead
of what actually appears, which is "[fixed]", like so:

 scatter-gather: off [fixed]

This patch corrects these errors and adds a unit test to prevent
regressions.

We also change the field name of `NICCapability.CanChange` to
`NICCapability.CanEnable` since this is more congruent with the existing
field `NICCapability.IsEnabled`. Changing the field name like this is
considered fine since anyone relying on the old field name was relying
on bad/buggy behaviour and we might as well hard-break those folks and
let them know about the bug...

Issue #111