fix Windows memory calculations
Win32_OperatingSystem.FreePhysicalMemory was incorrectly being used for
"usable bytes". Usable bytes refers to the amount of physical RAM that
is available to the operating system for use (as opposed to the total
amount of physical RAM within DIMMs attached to the system). In WMI, the
Win32_OperatingSystem.TotalVisibleMemorySize is the actual *usable*
memory (in KB) that the operating system has available to it.

The total *physical* RAM in DIMMs attached to the system can be
calculated by looking at the sum of the memory modules attached to the
system's SizeBytes field.

Fixes Issue #176