ghw-snapshot: capture cpu, memory, topology data
Add support to capture data ghw cares about in
/sys/system/devices/{cpu,node}.
With this patch, we can create snapshot we can reuse
to test the `cpu`, `memory` and `topology` packages using
the codeto consume snapshot added in commit 1a9fae7

There is quite a lot of opportunity to extract code
from ghw-snapshot and add it to the newl created `snapshot`
pkg, to make it easier to reuse and more testable.

This refactoring is deferred to a later change because
the new logic added is complex enough to stand in its own.

Signed-off-by: Francesco Romani <fromani@redhat.com>