snapshot: clonetree: handle PCI bridges
Modern (x86_64) systems have multiple PCI buses.
For ghw needs, considering the abstraction provided by the linux kernel,
there is no practical difference from PCI-express and PCI to date,
so we will call them all "PCI".

PCI-to-PCI (see initial note about PCI express) are common, and their
topology depends on the CPU, mainboard and system manufacturer choice.

The linux kernel does not flatten the PCI tree in sysfs.
This choice provides the closest representation to reality, and
unfortunately is also easy to overlook
when exploring simple PCI device trees, such as ones found on laptops,
but becomes evident when exploring more complex trees, like the ones
we can see on servers, even low end.

This means, in case of PCI bridges, we will have
entries inside the directory which represent a bus which
will be both devices and root of subtrees.

In other words, these entries will contain BOTH subdirectories
to represent the devices attached to that bridge, and the files
which describe the bridge device per se.

For ghw, the net effect is we cannot just scan the entries
in `/sys/bus/pci/devices` and be done with it; we need to detect
the entries which represent PCI bridges, and scan those recursively.
Otherwise, we miss to gather the data pertaining to these devices.

So we need a smarter function that does this traversal, moving away
from the nice static list of paths we previously had.

Without this patch (example taken from dev laptop):
{host bridge}
-[0000:00]-+-00.0
           +-02.0
           +-14.0
           +-14.2
           +-16.0
           +-16.3
           +-1c.0 {PCI bridge}
           +-1c.2 {PCI bridge}
           +-1d.0 {PCI bridge}
           +-1f.0
           +-1f.2
           +-1f.3
           +-1f.4
           \-1f.6

With this patch (example taken from dev laptop):
{host bridge}
-[0000:00]-+-00.0
           +-02.0
           +-14.0
           +-14.2
           +-16.0
           +-16.3
           +-1c.0 | -+-3a {PCI bridge}
                     \-00.0 ** PREVIOUSLY NOT DETECTED!
           +-1c.2 | -+-3c {PCI bridge}
                     \-00.0 ** PREVIOUSLY NOT DETECTED!
           +-1d.0 {PCI bridge - no devices here}
           +-1f.0
           +-1f.2
           +-1f.3
           +-1f.4
           \-1f.6

Signed-off-by: Francesco Romani <fromani@redhat.com>