pci: don't pass chroot to pcidb with snapshots
Snapshot implementation leverages chroot and transparently set it
to the unpacked snapshot root. This is the intentional behaviour
and ghw supports it.
But the `pci` package passes throu the chroot settings to `pcidb`
and the latter doesn't know or care about ghw snapshots - and rightly
so!

hence, we need to pass through the `chroot` setting only if explicitely
set by client code - so if no snapshot settings are present.

Signed-off-by: Francesco Romani <fromani@redhat.com>