rename PCIAddress.Slot to PCIAddress.Device
As noted by @pearsonk in #220, the PCIAddress.Slot field was a bit of a
misnomer, since the more common representation of a PCIAddress is
Bus:Device.Function, not Bus:Slot.Function. Actual PCI slot information
may be retrieved from DMI, but what we were referring to was not the
actual PCI slot information but rather the PCI address device component.

Fixes Issue #220