pci: remove redundant multiple pciaddr parsing
We collected quite few helpers over time to add useful extra
info to the `pci.Device`. This leads to duplicated (and useless)
parsing of the `address string`.
Since these helpers are never meant to be called outside the pci
package, let's do the parsing just once and let's pass through
the `pciaddr.Address`.

This saves some (real) cpu cycles and yields
to better error message in the error path.

Signed-off-by: Francesco Romani <fromani@redhat.com>