memory: topology: expose per-numa memory
This patch wants to add memory informations on topology.Node,
so we can have per-NUMA-zone memory information.
In order to do so, we extract a `memory.Area` struct from `memory.Info`
to describe the basic properties of a memory zone, and
we add it to `topology.Node`.
In the future, we may want to add (or move) more fields
to this basic building block, and we will get them for free
in both system-wide and per-numa memory info.

Fixes: https://github.com/jaypipes/ghw/issues/200
Signed-off-by: Francesco Romani <fromani@redhat.com>