snapshot: actually copy the 'driver' entry
In commits 3500903a6e and f15e79ca87 we added support to copy
the "driver" entries from sysfs to a snapshot.
Due to a rebase glitch, we actually missed to merge the last
bit which enables the actual copy.

This bug is not evident until a new snapshot is created,
and this is why it wasn't noticed up until now.

Signed-off-by: Francesco Romani <fromani@redhat.com>