snapshot: fix pack walk
The documentation of the callback (WalkFunc) of
`filepath.Walk` reads like:

  The err argument reports an error related to path, signaling that Walk
  will not walk into that directory. The function can decide how to
  handle that error; as described earlier, returning the error will cause
  Walk to stop walking the entire tree.

Here we want to always keep walking, and this is why we ignored
the argument. We still want to do this, so let's rename the
argument to signal the intent of the code, and to avoid unwanted
shadowing with the local variable.

Signed-off-by: Francesco Romani <fromani@redhat.com>