Add WithLoopDevices to report loop devices
Currently we skipp over all loop devices in the system and with no
possibility of reporting those.

This patch introduces a new option to Block that allows us to report
loop devices as disk and their partitions like if they were normal
disks.

This patch does not modify any existing functionality or alters anything
with the default values.

With WithLoopDevices enabled this reports the same info for loop devices
as with other disks, including size, fs, mountpoints and so on.

Signed-off-by: Itxaka <igarcia@suse.com>