bring in github.com/jaypipes/pcidb@v1.0.0
I cut a new release of the pcidb library (v1.0.0) to address a
particular issue (jaypipes/pcidb#28) that was causing pain for the
maintainers of the PCI-IDS database. This PR simply brings in that
latest patched version of pcidb, which includes disabling by default the
network fetch of the PCI-IDS database when no local database file is
found.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>