Fix wrong fs label as partition label
The label entry in the Partition struct should correspond to the
partition label, but we were getting the FS label instead.

This patch fixes that by getting the actual partition label for the
label value and adding an extra field on the Partition struct to add the
fs label, so they are separated and clear on which is which.

Signed-off-by: Itxaka <igarcia@suse.com>