supply override PCIDB_PATH env var in unit tests
Some of the unit tests inspect snapshot filesystems and for the GPU
Info-collection, the `github.com/jaypipes/pcidb` library is used to get
PCI information. That library no longer (as of version v0.9) does an
automatic fetch of the canonical pci.ids file if it cannot find a
pci.ids file in the local filesystem. It no longer does that network
fetch because the upstream maintainer of the pci.ids database noticed
too many fetches coming from projects that imported `ghw` which
transitively imported `pcidb`.

Because of this change in behaviour of no longer doing an automatic
network fetch of the pci.ids database, a unit test in `ghw` that was
relying on the pcidb library was no longer working. This patch fixes two
things related to the PCI handling in the `pkg/gpu` package. First, it
cleans up the parsing of the `/sys/class/drm` symlink contents by
looking through the path parts and attempting a regex match for a PCI
address. Second, it adds an env-var override to the GPU unit tests to
point `github.com/jaypipes/pcidb` to a specific pci.ids file that I've
included in the `testdata/` directory.

Closes Issue #375

Signed-off-by: Jay Pipes <jaypipes@gmail.com>